Use uintptr_t to cast a pointer to an integer to avoid compiler warnings on processors where sizeof(void *) > sizeof(int).
Former-commit-id: 5a006948336eec5e3e5a7ffde604c0851c29a017

+++ usr.sbin/ppp/chap.h
@@ -152,7 +152,7 @@ parameter_list|(
 name|a
 parameter_list|)
 define|\
-value|((struct chap *)((char *)a - (int)&((struct chap *)0)->auth))
+value|((struct chap *)((char *)a - (uintptr_t)&((struct chap *)0)->auth))
 end_define
 
 begin_struct
